[mini] echo: limit the number of digits
The previous version made printing [shift-in]0, for example, very difficult as:
- no \letter for shift-in
- \00170 would be interpreted as \0170 = x
- \017\060 would personally piss me off

Plus that's closer to the coreutils manpage anyway, so that's that.

BTW, I decided the POSIX spec and coreutils manpage would be what I'd try to
mimic under reasonable feasibility. So no Texinfo specs for me, thanks.
And I won't implement --help/--version/etc., that's terrible.

The doc belongs in the manpage!
Why don't we have a standardized ELF section to store versioning and/or
copyright and/or licensing metadata? 1970, take your non-standardized and
deprecated behaviours with you please.